perf(moe): pass int32 top-k IDs to FlashInfer MegaMoE
Remove the explicit int64 conversion before W4A8 fused staging and preserve int32 IDs for empty ranks. Add adapter regression coverage for dtype and zero-copy forwarding.

+++ python/sglang/srt/layers/moe/mega_moe.py
@@ -298,9 +298,9 @@ def run_mega_routed_experts(
             experts,
             hidden_states,
             topk_ids=(
-                topk_ids.to(torch.int64)
+                topk_ids
                 if topk_ids is not None
-                else hidden_states.new_empty((0, top_k), dtype=torch.int64)
+                else hidden_states.new_empty((0, top_k), dtype=torch.int32)
             ),
             topk_weights=(
                 topk_weights.to(torch.float32)
